Seperate object generate function for reuse
